+++ crates/miden-tx/src/executor/mod.rs
@@ -309,7 +309,6 @@ where
             script_mast_store,
             acct_procedure_index_map,
             self.authenticator,
-            tx_inputs.block_header().fee_parameters(),
             self.source_manager.clone(),
         );